requirements.txt
anyio==4.6.2.post1
ariadne==0.23.0
autopep8==2.3.1
aws-lambda-powertools==3.1.0
aws-xray-sdk==2.14.0
boto3==1.35.43
boto3-stubs==1.35.43
botocore==1.35.43
botocore-stubs==1.35.43
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
graphql-core==3.2.5
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
mypy-boto3-apigatewaymanagementapi==1.35.0
mypy-boto3-cognito-idp==1.35.18
mypy-boto3-dynamodb==1.35.24
pycodestyle==2.12.1
pycparser==2.22
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
types-awscrt==0.22.0
types-s3transfer==0.10.3
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
